* master:
  WebSession: fix for merge_usera_into_userb()

+++ modules/websession/lib/webuser.py
@@ -514,8 +514,11 @@ def merge_usera_into_userb(id_usera, id_userb):
     preferencea.update(preferenceb)
     set_user_preferences(id_userb, preferencea)
     try:
-        for table, dummy in CFG_WEBUSER_USER_TABLES:
-            run_sql("LOCK TABLE %s WRITE" % table)
+        ## FIXME: for the time being, let's disable locking
+        ## until we will move to InnoDB and we will have
+        ## real transitions
+        #for table, dummy in CFG_WEBUSER_USER_TABLES:
+            #run_sql("LOCK TABLE %s WRITE" % table)
         index = 0
         table = ''
         try:
@@ -531,7 +534,9 @@ def merge_usera_into_userb(id_usera, id_userb):
             register_exception(alert_admin=True, prefix=msg)
             raise
     finally:
-        run_sql("UNLOCK TABLES")
+        ## FIXME: locking disabled
+        #run_sql("UNLOCK TABLES")
+        pass
 
 def loginUser(req, p_un, p_pw, login_method):
     """It is a first simple version for the authentication of user. It returns the id of the user,
@@ -571,7 +576,7 @@ def loginUser(req, p_un, p_pw, login_method):
             if res:
                 ## User was already registered with this external method.
                 id_user = res[0][0]
-                old_email = run_sql("SELECT email FROM user WHERE id=%s", (id_user,))[0][0]
+                old_email = run_sql("SELECT email FROM user WHERE id=%s", (id_user,))[0][0].lower()
                 if old_email != p_email:
                     ## User has changed email of reference.
                     res = run_sql("SELECT id FROM user WHERE email=%s", (p_email,))
@@ -579,6 +584,8 @@ def loginUser(req, p_un, p_pw, login_method):
                         ## User was also registered with the other email.
                         ## We should merge the two!
                         new_id = res[0][0]
+                        if new_id == id_user:
+                            raise AssertionError("We should not reach this situation: new_id=%s, id_user=%s, old_email=%s, p_email=%s" % (new_id, id_user, old_email, p_email))
                         merge_usera_into_userb(id_user, new_id)
                         run_sql("DELETE FROM user WHERE id=%s", (id_user, ))
                         for row in run_sql("SELECT method FROM userEXT WHERE id_user=%s", (id_user, )):
